API de fotos completadas La validación de los tacos de la descripción nos la hemos llevado al modelo

+++ frikr/photos/api.py
@@ -6,11 +6,16 @@ from rest_framework.response import Response
 from django.shortcuts import get_object_or_404
 from rest_framework import status
 from models import Photo
-from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView
+from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView, GenericAPIView
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from permissions import UserPermission
+from django.db.models import Q
 
 
 
 class UserListAPI(APIView):
+
+    permission_classes = (UserPermission,)
 
     def get(self, request):
         """
@@ -35,57 +40,89 @@ class UserListAPI(APIView):
 
 class UserDetailAPI(APIView):
 
+
     def get(self, request, pk):
         """
         Devuelve el detalle de un usuario
         """
-        user = get_object_or_404(User, pk=pk) # recuperamos el User por pk si no encuentra lanza un 404
-        serializer = UserSerializer(user)
-        return Response(serializer.data)
-
+        user = get_object_or_404(User, pk=pk)  # recuperamos el User por pk si no encuentra lanza un 404
+        if request.user.is_superuser or request.user == user:
+            serializer = UserSerializer(user)
+            return Response(serializer.data)
+        else:
+            return Response(status=status.HTTP_404_NOT_FOUND)
 
     def put(self, request, pk):
         """
         Actualiza un usuario
         """
         user = get_object_or_404(User, pk=pk)
-        serializer = UserSerializer(user, data=request.DATA)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_202_ACCEPTED) # 202 ACCEPTED
+        if request.user.is_superuser or request.user == user:
+            serializer = UserSerializer(user, data=request.DATA)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_202_ACCEPTED) # 202 ACCEPTED
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST) # 400 BAD REQUEST
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST) # 400 BAD REQUEST
-
+            return Response(status=status.HTTP_404_NOT_FOUND)
 
     def delete(self, request, pk):
         """
         Elimina un usuario
         """
         user = get_object_or_404(User, pk=pk)
-        user.delete()
-        return Response({}, status=status.HTTP_204_NO_CONTENT) # 204 NO CONTENT
+        if request.user.is_superuser or request.user == user:
+            user.delete()
+            return Response({}, status=status.HTTP_204_NO_CONTENT) # 204 NO CONTENT
+        else :
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+
+class PhotoAPIQueryset:
+
+    def get_queryset(self) :
+        """
+        Si es superusuario, accede a todo
+        Si no es superusuario y está autentciado, accede a lo suyo y a lo público del resto
+        Si no está autenticado, sólo ve lo público
+        """
+        if self.request.user.is_superuser :
+            return Photo.objects.all()
+        elif self.request.user.is_authenticated() :
+            return Photo.objects.filter(Q(owner=self.request.user) | Q(visibility=Photo.VISIBILITY_PUBLIC))
+        else :
+            return Photo.objects.filter(visibility=Photo.VISIBILITY_PUBLIC)
 
 
 
-
-class PhotoListAPI(ListCreateAPIView):
+class PhotoListAPI(PhotoAPIQueryset, ListCreateAPIView):
     """
     API con endpoints de listado y creación de fotos
     """
     queryset = Photo.objects.all()
     serializer_class = PhotoListSerializer
+    permission_classes = (IsAuthenticatedOrReadOnly,)
 
 
     def get_serializer_class(self):
         return PhotoSerializer if self.request.method == "POST" else PhotoListSerializer
 
 
-class PhotoDetailAPI(RetrieveUpdateDestroyAPIView):
+    def pre_save(self, obj):
+        """
+        Asigna el autor de la foto antes de ser creada
+        """
+        obj.owner = self.request.user
+
+
+class PhotoDetailAPI(PhotoAPIQueryset, RetrieveUpdateDestroyAPIView):
     """
     API con endpoints de detalle, actualización y borrado
     """
     queryset = Photo.objects.all()
     serializer_class = PhotoSerializer
+    permission_classes = (IsAuthenticatedOrReadOnly,)
